requirements.txt
elevenlabs
python-dotenv
pandas
sqlalchemy
psycopg2-binary
